Updated update functions to make it look cleaner.
I aim to make this the cleanest looking OS there is, not the best. This
OS is intended for use, not for fun. If you don't like that, then I
don't mind. Just don't nag. That is why there are so many updates
happening. Once the desktop file comes out and the login system is
complete, there is NO way that I will release late. The OS should be out
within 30 minutes to an hour after completion.