Fixed (Again) CraftOS breaking bug
It turns out, that I broke my own code. If you call .redos/os/desktop
directly, the W and H variables for term.getSize() don't exist. If you
call it with the startup file (If you have one), the RedOS program
leaves the W and H variables when it closes, causing a potential stack
overflow and letting the desktop program have it's W and H variables.